requirements.txt
python-telegram-bot==13.7
faunadb==4.1.0
pytz==2021.1
